+++ apis/r/R/ReadIter.R
@@ -30,16 +30,13 @@ ReadIter <- R6::R6Class(
     #' @return \code{NULL} or one of arrow::\link[arrow]{Table}, \link{matrixZeroBasedView}
     read_next = function() {
       if (is.null(private$soma_reader_pointer)) {
-          NULL
-      } else {
-          if (self$read_complete()) {
-              warning("Iteration complete, returning NULL")
-              NULL
-          } else {
-              rl <- sr_next(private$soma_reader_pointer)
-              return(private$soma_reader_transform(rl))
-          }
+          return(NULL)
       }
+      if (self$read_complete()) {
+        return(private$.readComplete())
+      }
+
+      return(private$.read_next())
     },
 
     #' @description  Concatenate remainder of iterator
@@ -47,7 +44,6 @@ ReadIter <- R6::R6Class(
     concat = function() {
       .NotYetImplemented()
     }
-
   ),
 
   private = list(
@@ -56,8 +52,24 @@ ReadIter <- R6::R6Class(
     soma_reader_pointer = NULL,
 
     # to be refined in derived classes
-    soma_reader_transform = function(x) {
-      .NotYetImplemented()
+    soma_reader_transform = function(x) .NotYetImplemented(),
+
+    # Internal `read_next()` to avoid `self$read_complete()` checks
+    .read_next = function() {
+      if (is.null(private$soma_reader_pointer)) {
+        return(NULL)
+      }
+      rl <- sr_next(private$soma_reader_pointer)
+      return(private$soma_reader_transform(rl))
+    },
+
+    # Throw a warning for read completion
+    .readComplete = function() {
+      warning(warningCondition(
+        "Iteration complete, returning NULL",
+        class = "iterationCompleteWarning"
+      ))
+      return(NULL)
     }
 
   )