+++ apis/r/R/utils-readerTransformers.R
@@ -11,6 +11,33 @@ soma_array_to_arrow_table <- function(x) {
   arrow::as_arrow_table(x)
 }
 
+soma_array_to_arrow_table_concat <- function(it) {
+  stopifnot("'it' must be a 'ReadIter' object" = inherits(it, 'ReadIter'))
+  tbl <- it$read_next()
+  while (!it$read_complete()) {
+    nxt <- it$read_next()
+    tbl <- arrow::concat_tables(tbl, nxt)
+  }
+  return(tbl)
+}
+
+soma_array_to_sparse_matrix_concat <- function(it, zero_based = FALSE) {
+  stopifnot(
+    "'it' must be a 'ReadIter' object" = inherits(it, 'ReadIter'),
+    "'zero_based' must be TRUE or FALSE" = isTRUE(zero_based) || isFALSE(zero_based)
+  )
+  mat <- it$read_next()
+  while (!it$read_complete()) {
+    mat <- if (isTRUE(zero_based)) {
+      mat$sum(it$read_next())
+    } else {
+      mat + it$read_next()
+    }
+  }
+  return(mat)
+}
+
+
 #' Transformer function: Arrow table to Matrix::sparseMatrix
 #'
 #' @description Converts a \link[arrow]{Table} of sparse format (columns: