+++ apis/r/man/roxygen/templates/param-repr-read.R
@@ -1,0 +1,5 @@
+#' @param repr Representation of the sparse matrix to return; choose from:
+#' \itemize{
+#'  \item \dQuote{\code{T}}: returns a
+#'   \code{\link[Matrix:TsparseMatrix-class]{TsparseMatrix}}
+#' }