+++ apis/r/NEWS.md
@@ -5,6 +5,7 @@
 * Add support for ingestion of `SeuratCommand` logs
 * Add support for outgestion of `SeuratCommand` logs
 * Add support for reading `*m` and `*p` layers from `SOMAExperimentAxisQuery`
+* Add support for blockwise iteration
 
 # 1.7.0